Fix `className` warning
In this commit, I've addressed and resolved a warning that was appearing in the console related to the use of the `className` attribute. The warning message, which stated "Received `false` for a non-boolean attribute `className`," has been fixed. The issue was resolved by ensuring that the `className` attribute receives a string value as intended, preventing any erroneous `false` values from being passed.